+++ crates/zfs-ondisk/src/lib.rs
@@ -21,6 +21,7 @@ pub mod label;
 pub mod nvlist;
 pub mod part;
 pub mod raidz;
+pub mod signature;
 pub mod skein;
 pub mod spacemap;
 pub mod uberblock;